+++ python/dartpy/simulation_experimental/module.cpp
@@ -1652,6 +1652,9 @@ void defSimulationExperimentalModule(nb::module_& m)
           "projected_newton_iterative_solves",
           &sim::DeformableSolverDiagnostics::projectedNewtonIterativeSolves)
       .def_ro(
+          "projected_newton_matrix_free_solves",
+          &sim::DeformableSolverDiagnostics::projectedNewtonMatrixFreeSolves)
+      .def_ro(
           "projected_newton_iterative_iterations",
           &sim::DeformableSolverDiagnostics::projectedNewtonIterativeIterations)
       .def_ro(
@@ -1691,7 +1694,10 @@ void defSimulationExperimentalModule(nb::module_& m)
           &sim::DeformableMaterialProperties::useAdaptiveBarrierStiffness)
       .def_rw(
           "use_iterative_linear_solver",
-          &sim::DeformableMaterialProperties::useIterativeLinearSolver);
+          &sim::DeformableMaterialProperties::useIterativeLinearSolver)
+      .def_rw(
+          "use_matrix_free_linear_solver",
+          &sim::DeformableMaterialProperties::useMatrixFreeLinearSolver);
 
   nb::class_<sim::DeformableEdge>(m, "DeformableEdge")
       .def(